Project_Level_3_Image_Caption_Generator/requirements.txt
Flask==2.3.3
tensorflow==2.15.0
Pillow==10.3.0
numpy==1.26.4
Werkzeug==2.3.8

absl-py==1.4.0
altair==5.1.0
astunparse==1.6.3
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
flatbuffers==23.5.26
fonttools==4.42.1
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.57.0
h5py==3.9.0
idna==3.4
importlib-metadata==6.8.0
jax==0.4.14
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
ml-dtypes==0.2.0
nltk==3.8.1
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
protobuf==4.24.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.2
rpds-py==0.10.0
rsa==4.9
scipy==1.11.2
six==1.16.0
smmap==5.0.0
streamlit==1.38.0
tenacity==8.2.3
tensorboard==2.15.2
tensorboard-data-server==0.7.1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.16
validators==0.21.2
watchdog==3.0.0
wrapt==1.14.1
zipp==3.16.2